refactor: rewrite problem page
